Revert "Refactored BaseApp into multiple components"
This reverts commit 093d0941aa27a60ef966e54ad33c2f0087c331a7.
